+++ app/build.gradle.kts
@@ -58,7 +58,7 @@ dependencies {
 
     // Advanced animations (optional - for future enhancements)
     implementation("com.airbnb.android:lottie-compose:6.3.0")
-    implementation(libs.androidx.navigation.compose.jvmstubs)
+    implementation("androidx.navigation:navigation-compose:2.7.7")
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
