+++ app/src/main/java/com/mkj/sooooorrrrryyyyy/ui/Routes.kt
@@ -2,8 +2,8 @@ package com.mkj.sooooorrrrryyyyy.ui
 
 sealed class Routes(val route: String) {
     object ChoosingScreen : Routes("ChooseApology")
-    object FriendInput : Routes("splash")
+//    object FriendInput : Routes("FriendInputScreen")
     object  FriendScreen: Routes("SorryScreen")
     object LoverInput : Routes("LoveInputScreenLove")
-    object  LoverScreen: Routes("LoveDisplayScreenLove")
+    object  LoverScreen: Routes("LoveApologyScreen")
 }
